+++ cypress/integration/StageEventListPage/StageEventListNavigation/index.js
@@ -3,11 +3,11 @@ beforeEach(() => {
 });
 
 Given('you are visiting Stage Event List Page by url', () => {
-    cy.visit('/#/enrollment/stageEvents?enrollmentId=wBU0RAsYjKE&stageId=A03MvHHogjR&orgUnitId=DiszpKrYNg8');
+    cy.visit('/#/enrollment/stageEvents?enrollmentId=wBU0RAsYjKE&stageId=A03MvHHogjR&orgUnitId=DiszpKrYNg8&programId=IpHINAT79UW');
 });
 
 Then('you should see the content of the page', () => {
     cy.get('[data-test="stage-event-list"]')
-        .contains('Stages and Events')
+        .contains('Birth')
         .should('exist');
 });