+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.component.js
@@ -2,10 +2,12 @@
 import React, { type ComponentType } from 'react';
 import withStyles from '@material-ui/core/styles/withStyles';
 import { spacersNum } from '@dhis2/ui';
+import i18n from '@dhis2/d2-i18n';
 import { WidgetStagesAndEvents } from '../../../WidgetStagesAndEvents';
 import { WidgetEnrollment } from '../../../WidgetEnrollment';
 import { WidgetProfile } from '../../../WidgetProfile';
 import type { Props, PlainProps } from './EnrollmentPageDefault.types';
+import { WidgetFeedback } from '../../../WidgetFeedback';
 
 const getStyles = ({ typography }) => ({
     columns: {
@@ -21,6 +23,9 @@ const getStyles = ({ typography }) => ({
         flexShrink: 1,
         paddingLeft: spacersNum.dp16,
         width: 360,
+        display: 'flex',
+        flexDirection: 'column',
+        gap: '12px',
     },
     title: {
         ...typography.title,
@@ -36,6 +41,8 @@ export const EnrollmentPageDefaultPlain = ({
     enrollmentId,
     stages,
     onDelete,
+    widgetEffects,
+    hideWidgets,
     classes,
 }: PlainProps) => (
     <>
@@ -48,6 +55,12 @@ export const EnrollmentPageDefaultPlain = ({
                 />
             </div>
             <div className={classes.rightColumn}>
+                {!hideWidgets.feedback && (
+                    <WidgetFeedback
+                        feedback={widgetEffects?.feedbacks}
+                        emptyText={i18n.t('No feedback for this enrollment yet')}
+                    />
+                )}
                 <WidgetProfile teiId={teiId} programId={program.id} />
                 <WidgetEnrollment
                     teiId={teiId}