+++ src/core_modules/capture-core/components/Pages/Enrollment/EnrollmentPageDefault/EnrollmentPageDefault.container.js
@@ -1,5 +1,5 @@
 // @flow
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
 import log from 'loglevel';
 import { errorCreator } from 'capture-core-utils';
 // $FlowFixMe
@@ -11,6 +11,7 @@ import { useEnrollmentsAndAttributes, useProgramMetadata, useProgramStages } fro
 import { runRulesForEnrollment } from './runRulesForEnrollment';
 import { urlArguments } from '../../../../utils/url';
 import { deleteEnrollment } from '../EnrollmentPage.actions';
+import { useFilteredWidgetData } from './hooks/useFilteredWidgetData';
 
 
 export const EnrollmentPageDefault = () => {
@@ -40,13 +41,27 @@ export const EnrollmentPageDefault = () => {
         log.error(errorCreator('Enrollment page could not be loaded')({ programMetaDataError, enrollmentsError }));
     }
 
-    const [, setRuleEffects] = useState(undefined);
+    const [ruleEffects, setRuleEffects] = useState(undefined);
+    const outputEffects = useFilteredWidgetData(ruleEffects);
     useEffect(() => {
         const effects = runRulesForEnrollment({ orgUnit, program, programMetadata, enrollment, attributes });
         if (effects) {
+            // $FlowFixMe
             setRuleEffects(effects);
         }
     }, [orgUnit, program, programMetadata, enrollment, attributes]);
+
+    const flatRuleActionLocations = useMemo(() => program.programRules.map(item => item.programRuleActions
+        .map(rule => rule.location || null))
+        .flat(), [program.programRules]);
+
+    const hideWidgets = useMemo(() => {
+        const hideWidgetObject = {};
+        hideWidgetObject.feedback = !flatRuleActionLocations.includes('feedback');
+        hideWidgetObject.indicator = !flatRuleActionLocations.includes('indicators');
+        return hideWidgetObject;
+    }, [flatRuleActionLocations]);
+
 
     const onDelete = () => {
         history.push(
@@ -63,6 +78,8 @@ export const EnrollmentPageDefault = () => {
             events={enrollment?.events ?? []}
             enrollmentId={enrollmentId}
             onDelete={onDelete}
+            widgetEffects={outputEffects}
+            hideWidgets={hideWidgets}
         />
     );
 };