+++ src/core_modules/capture-core/metaData/index.js
@@ -24,6 +24,7 @@ export { SearchGroup } from './SearchGroup';
 export { InputSearchGroup } from './InputSearchGroup';
 export { TrackedEntityType, TeiRegistration } from './TrackedEntityType';
 export { SystemSettings } from './SystemSettings';
+
 // helpers
 export {
     getScopeFromScopeId,